sửa thuộc tính các file từ decimal về double

+++ Bài Tập Lớn/BLL/ChiTietPhienBLL.cs
@@ -29,7 +29,7 @@ namespace Bài_Tập_Lớn.BLL
             return chiTietPhienDAL.TimTheoMaPhien(maPhien);
         }
 
-        public decimal TinhTongTienTheoPhien(string maPhien)
+        public double TinhTongTienTheoPhien(string maPhien)
         {
             if (string.IsNullOrWhiteSpace(maPhien))
             {
